Move _transfer from MeshSeq to Solver
